[Tools] - New tool plug in for the editor that will break a texture atlas down to smaller images using the sprites assigned to the atlas. [Tools] - Fixed issue where memory was not being freed in texture atlas tool when it should have been. [Tools] - Added tool for creating texture atlas from individual images [Editor] - Fixed a bug in the file system folder browser where the current directory information was incorrect when creating new directories. [Editor] - Fixed an issue where deleting or renaming a folder from the folder browser would not work due to an incorrect selection state.

+++ Tools/Editor/Gorgon.Editor.API/Rendering/DefaultToolRenderer.cs
@@ -329,7 +329,7 @@ namespace Gorgon.Editor.Rendering
         /// </remarks>
         protected virtual void OnRenderBackground()
         {
-            var textureSize = new DX.RectangleF(0, 0, ClientSize.Width / BackgroundPattern.Width, ClientSize.Height / BackgroundPattern.Height);
+            var textureSize = new DX.RectangleF(0, 0, ClientSize.Width / (float)BackgroundPattern.Width, ClientSize.Height / (float)BackgroundPattern.Height);
 
             Renderer.Begin();
             Renderer.DrawFilledRectangle(new DX.RectangleF(0, 0, ClientSize.Width, ClientSize.Height), GorgonColor.White, BackgroundPattern, textureSize);
